docs: raise awareness of resolver used inside workspace
Workspaces by default use the resolver in version 1. It's been
some time already since the 2021 edition which made version 2
a default for the alone packages, but yet most of the projects
that make a use of the workspaces still depends on an old resolver.

By being explicit about the resolver tag inside the workspace
we can lower the usage of older resolver for the new projects.